+++ app/main.py
@@ -17,13 +17,9 @@ from app.routers.study import router as study_router
 from app.routers.analysis import router as analysis_router
 from app.routers.search import router as search_router
 from app.routers.recommend import router as recommend_router
-from app.routers.coach import router as coach_router
 from app.routers.teacher import router as teacher_router
 from app.routers.stats import router as stats_router
-from app.routers.achievements import router as achievements_router
-from app.routers.transfer import router as transfer_router
 from app.routers.ops import router as ops_router
-from app.routers.features import router as features_router
 
 setup_logging()
 
@@ -33,8 +29,8 @@ app.add_middleware(RateLimitMiddleware, max_requests=rate_limit_per_min())
 app.add_middleware(
     CORSMiddleware,
     allow_origins=["*"],
-    allow_methods=["POST"],
-    allow_headers=["Content-Type"],
+    allow_methods=["*"],
+    allow_headers=["*"],
 )
 app.include_router(health_router)
 app.include_router(words_router)
@@ -46,13 +42,9 @@ app.include_router(study_router)
 app.include_router(analysis_router)
 app.include_router(search_router)
 app.include_router(recommend_router)
-app.include_router(coach_router)
 app.include_router(teacher_router)
 app.include_router(stats_router)
-app.include_router(achievements_router)
-app.include_router(transfer_router)
 app.include_router(ops_router)
-app.include_router(features_router)
 
 
 @app.on_event("startup")